[refactor] ChipInput 로직 분리 - ChipInput 로직 분리
